crud/requirements.txt
django
pillow
whitenoise
djangorestframework
django-cors-headers ##for making api call to the front end